ai_server/requirements.txt
fastapi
uvicorn
sqlalchemy
pillow
opencv-python
numpy
RealESRGAN
pydantic-settings
psycopg2-binary
asyncpg
greenlet
torch
torchvision
python-multipart
python-dotenv
alembic
gunicorn
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
slowapi==0.1.9
